backend/requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.34.0
python-multipart>=0.0.18
pydantic-settings>=2.7.0
google-generativeai>=0.5.0
numpy>=2.2.0
scikit-learn>=1.6.0
hdbscan>=0.8.40
midiutil>=1.2.1
pandas>=2.2.0
python-dotenv>=1.0.1
